lat long

+++ lib/models/weather_model.dart
@@ -40,10 +40,10 @@ class Current {
   double windKph;
   int windDegree;
   String windDir;
-  int pressureMb;
+  double pressureMb;
   double pressureIn;
-  int precipMm;
-  int precipIn;
+  double precipMm;
+  double precipIn;
   int humidity;
   int cloud;
   double feelslikeC;
@@ -54,9 +54,9 @@ class Current {
   double heatindexF;
   double dewpointC;
   double dewpointF;
-  int visKm;
-  int visMiles;
-  int uv;
+  double visKm;
+  double visMiles;
+  double uv;
   double gustMph;
   double gustKph;
